+++ Cogworks.AzureSearch.sln
@@ -13,17 +13,17 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UnitTests", "UnitTests", "{
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.UnitTests\Cogworks.AzureSearch.UnitTests.csproj", "{37898883-18CD-47A7-BA59-E4FC1FD8B8E1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cogworks.AzureSearch.UmbracoIoc.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.UmbracoIoc.UnitTests\Cogworks.AzureSearch.UmbracoIoc.UnitTests.csproj", "{31032AAC-7779-4DEF-81D4-B6143A716C4F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cogworks.AzureSearch.IoC.Umbraco.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.IoC.Umbraco.UnitTests\Cogworks.AzureSearch.IoC.Umbraco.UnitTests.csproj", "{31032AAC-7779-4DEF-81D4-B6143A716C4F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.IoC.LightInject", "src\Cogworks.AzureSearch.IoC.LightInject\Cogworks.AzureSearch.IoC.LightInject.csproj", "{4C606328-9108-4584-8E4A-04450B9A1170}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.IoC.Microsoft", "src\Cogworks.AzureSearch.IoC.Microsoft\Cogworks.AzureSearch.IoC.Microsoft.csproj", "{E60BD2A0-C5ED-4EFE-8F8A-15FB26AD69B4}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.MicrosoftIoc.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.MicrosoftIoc.UnitTests\Cogworks.AzureSearch.MicrosoftIoc.UnitTests.csproj", "{1F9F4E09-E496-49F2-A1EB-D4EB3AC3E721}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.IoC.Microsoft.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.IoC.Microsoft.UnitTests\Cogworks.AzureSearch.IoC.Microsoft.UnitTests.csproj", "{1F9F4E09-E496-49F2-A1EB-D4EB3AC3E721}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.AutofacIoc.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.AutofacIoc.UnitTests\Cogworks.AzureSearch.AutofacIoc.UnitTests.csproj", "{0648AF35-71D5-4BA4-A1BF-0476AAC71BEE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.IoC.Autofac.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.IoC.Autofac.UnitTests\Cogworks.AzureSearch.IoC.Autofac.UnitTests.csproj", "{0648AF35-71D5-4BA4-A1BF-0476AAC71BEE}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.LightInject.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.LightInject.UnitTests\Cogworks.AzureSearch.LightInject.UnitTests.csproj", "{90E65F2C-9592-473C-83BB-FC109A11C47D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch.IoC.LightInject.UnitTests", "tests\UnitTests\Cogworks.AzureSearch.IoC.LightInject.UnitTests\Cogworks.AzureSearch.IoC.LightInject.UnitTests.csproj", "{90E65F2C-9592-473C-83BB-FC109A11C47D}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Cogworks.AzureSearch", "src\Cogworks.AzureSearch\Cogworks.AzureSearch.csproj", "{94F77E26-4E02-4ED5-BD54-45E3AAA70090}"
 EndProject
