+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Autofac.UnitTests/Cogworks.AzureSearch.IoC.Autofac.UnitTests.csproj
