+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Autofac.UnitTests/Models/ThirdTestDocumentModel.cs
@@ -1,9 +1,9 @@
 ﻿using Azure.Search.Documents.Indexes;
 using Cogworks.AzureSearch.Models;
 
-namespace Cogworks.AzureSearch.AutofacIoc.UnitTests.Models
+namespace Cogworks.AzureSearch.IoC.Autofac.UnitTests.Models
 {
-    public class NotRegisteredTestDocumentModel : IModel
+    public class ThirdTestDocumentModel : IModel
     {
         [SimpleField(IsKey = true, IsFilterable = true)]
         [SearchableField()]
