+++ tests/UnitTests/Cogworks.AzureSearch.IoC.LightInject.UnitTests/Cogworks.AzureSearch.IoC.LightInject.UnitTests.csproj
