+++ tests/UnitTests/Cogworks.AzureSearch.IoC.LightInject.UnitTests/Models/NotRegistredTestDocumentModel.cs
@@ -1,9 +1,9 @@
 ﻿using Azure.Search.Documents.Indexes;
 using Cogworks.AzureSearch.Models;
 
-namespace Cogworks.AzureSearch.LightInject.UnitTests.Models
+namespace Cogworks.AzureSearch.IoC.LightInject.UnitTests.Models
 {
-    public class FirstTestDocumentModel : IModel
+    public class NotRegisteredTestDocumentModel : IModel
     {
         [SimpleField(IsKey = true, IsFilterable = true)]
         [SearchableField()]
