+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Microsoft.UnitTests/Cogworks.AzureSearch.IoC.Microsoft.UnitTests.csproj
