+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/Models/FirstTestDocumentModel.cs
