+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/Models/NotRegistredTestDocumentModel.cs
