+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/Models/SecondTestDocumentModel.cs
