+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/Models/ThirdTestDocumentModel.cs
