+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/Properties/AssemblyInfo.cs
