+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/Searchers/CustomTestSearch.cs
