+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/UmbracoIocExtensionTests.cs
