+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/app.config
