+++ tests/UnitTests/Cogworks.AzureSearch.IoC.Umbraco.UnitTests/packages.config
